separated utils function due to circular includes
